rename file to CustomButtons.tsx to reflect its variety of helper button functions and update imports
